Reorganized, View Spawn, Building Manager
Reorganized some of the files to be within their respective Attempt folders
Fixed the Scroll View Spawning script to have the correct placement. Still need to get the scrolling part figured out.
Added Building Manager and Building scripts to keep track of the timer, quantity made per tic, production icon, and every time it successfully produces an item it sends it directly to the manager as a serialized class